update 코드정리
- 중복으로 쓰이는 이메일, 닉네임 중복 service단으로 이동
- SecurityConfig permitAll 임시추가
